college-management-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib
email-validator